+++ src/providers/index.ts
@@ -1,4 +1,3 @@
-export * from "./consumerVirtualTextDocumentProvider";
 export * from "./consumerTableViewProvider";
 export * from "./consumerSession";
 export * from "./outputChannelProvider";